@@ -1,15 +1,14 @@
-import { Component, OnInit } from "@angular/core";
-import { debounceTime } from "rxjs/operators";
+import { Component, OnDestroy, OnInit } from "@angular/core";
+import { debounceTime, switchMap, takeUntil, tap } from "rxjs/operators";
 import { FormBuilder, FormControl, FormGroup, ValidationErrors, Validators } from "@angular/forms";
 import { ApiService } from "../../../services/api/api.service";
-import { NzModalService } from "ng-zorro-antd";
+import { NzModalService } from "ng-zorro-antd/modal";
 import { TranslateService } from "@ngx-translate/core";
-import { UserLogged } from "@app/models";
+import { Organization, UserLogged } from "@app/models";
 import { ECOE } from "../../../models";
 import { UserService } from "@app/services/user/user.service";
-
 import { Router } from "@angular/router";
-import { Observable, Observer } from "rxjs";
+import { Observable, Observer, ReplaySubject, Subscription, defer, from } from "rxjs";
 import { SharedService } from "@app/services/shared/shared.service";
 
 @Component({
@@ -17,19 +16,21 @@ import { SharedService } from "@app/services/shared/shared.service";
   templateUrl: "./home.component.html",
   styleUrls: ["./home.component.less"],
 })
-export class HomeComponent implements OnInit {
+export class HomeComponent implements OnInit, OnDestroy {
   showCreateEcoe: boolean;
   ecoesList: ECOE[];
   ecoeForm: FormControl;
   ecoesDelist: ECOE[];
   ecoe: ECOE;
-  organization: any;
+  organization: Organization;
 
   user: UserLogged;
   Listed: any;
   Delisted: any;
 
   validateForm!: FormGroup;
+
+  private destroyed$: ReplaySubject<boolean> = new ReplaySubject(1);
 
   constructor(
     private formBuilder: FormBuilder,
@@ -44,23 +45,42 @@ export class HomeComponent implements OnInit {
       ecoeName: ['', [Validators.required], [this.userNameAsyncValidator]],
     });
   }
+  ngOnDestroy(): void {
+    this.destroyed$.next(true);
+    this.destroyed$.complete();
+  }
+
 
   ngOnInit() {
     this.Listed = true;
     this.Delisted = false;
     this.ecoeForm = this.formBuilder.control("", Validators.required);
-    this.loadEcoes();
+
+    if (this.userService.userData) {
+      // Cuando le das a la flecha para atrás
+      // se seguiran viendo las ecoes
+      this.user = this.userService.userData;
+      this.organization = this.user.user.organization;
+      this.loadEcoes();
+    } else {
+      this.userService.userDataChange.pipe(takeUntil(this.destroyed$)).subscribe((user) => {
+        this.user = user;
+        this.organization = this.user.user.organization;
+        this.loadEcoes();
+      });
+    }
+  }
+
+  async loadEcoes(): Promise<void> {
+    this.ecoesList = await ECOE.query({
+      where: {organization: this.user.user.organization},
+      sort: {$uri: false}
+    }) as ECOE[];
   }
 
   closeDrawer() {
     this.showCreateEcoe = false;
     this.ecoeForm.reset();
-  }
-
-  async loadEcoes() {
-    ECOE.query<ECOE>().then((_ecoes) => {
-      this.ecoesList = _ecoes;
-    });
   }
 
   showListed() {
@@ -88,7 +108,9 @@ export class HomeComponent implements OnInit {
       restore: "restore",
     };
     const resource = "ecoes/archive/" + id + "/restore";
-    this.apiService.createResource(resource, body).subscribe((result) => {
+    this.apiService.createResource(resource, body).
+      pipe(takeUntil(this.destroyed$))
+      .subscribe((result) => {
       if (result) {
         ECOE.dearchive();
         window.location.reload();
