@@ -1,8 +1,10 @@
-import { Component, OnInit, CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
+import { Component, OnInit, CUSTOM_ELEMENTS_SCHEMA, OnDestroy } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
 import { ApiService } from "@app/services/api/api.service";
 import { ECOE, Job, Student, Round, Shift } from "@models/index";
 import { Item, Pagination } from "@openecoe/potion-client";
+import { ReplaySubject } from "rxjs";
+import { takeUntil } from "rxjs/operators";
 
 interface ISummaryItems {
   total: number;
@@ -14,8 +16,9 @@ interface ISummaryItems {
   templateUrl: "./ecoe-results.component.html",
   styleUrls: ["./ecoe-results.component.less"],
 })
-export class EcoeResultsComponent implements OnInit {
-  areGenerated: any;
+export class EcoeResultsComponent implements OnInit, OnDestroy {
+  areGeneratedReport: boolean;
+  areGeneratedCSV: boolean;
   ecoeId: number;
   completion: number;
   completion_csv: number;
@@ -45,6 +48,8 @@ export class EcoeResultsComponent implements OnInit {
   stages: ISummaryItems = { total: 0, show: false, loading: true };
   ecoeName: any;
 
+  private destroyed$: ReplaySubject<boolean> = new ReplaySubject(1);
+  
   constructor(
     private api: ApiService,
     private route: ActivatedRoute,
@@ -102,6 +107,12 @@ export class EcoeResultsComponent implements OnInit {
       return;
     });
   }
+
+  ngOnDestroy(): void {
+    this.destroyed$.next(true);
+    this.destroyed$.complete();
+  }
+  
   onBack() {
     this.router.navigate(["/ecoe"]).finally();
   }
@@ -110,15 +121,16 @@ export class EcoeResultsComponent implements OnInit {
     Job.fetch<Job>(this.job_id_reports, { cache: false }).then(() => {
       this.progress = this.ecoe_job_reports.progress;
       this.completion = this.progress;
-      this.file_name_reports = this.ecoe_job_reports.file;
     });
+
     var completation = setInterval(() => {
       Job.fetch<Job>(this.job_id_reports, { cache: false }).then(() => {
         this.progress = this.ecoe_job_reports.progress;
         if (this.progress == 100.0) {
           clearInterval(completation);
           this.job_reports_file = this.ecoe_job_reports.uri;
-          this.areGenerated == true;
+          this.areGeneratedReport = true;
+          this.file_name_reports = this.ecoe_job_reports.file;
         }
       });
     }, 500);
@@ -128,18 +140,18 @@ export class EcoeResultsComponent implements OnInit {
     this.api.getJobFile(this.job_id_reports, this.file_name_reports);
   }
 
-  async generateEcoeData() {
+  generateEcoeData() {
     this.progress_csv = 0;
     this.btn_csv = false;
     this.btn_dwl_csv = false;
-    await this.api
+    this.api
       .postResource("ecoes/" + this.ecoeId + "/csv")
-      .subscribe((value) => value);
-    this.updateProgressCsv();
+      .pipe(takeUntil(this.destroyed$))
+      .subscribe(() => this.updateProgressCsv());
   }
   async updateProgressCsv() {
     await ECOE.fetch<ECOE>(this.ecoeId, { cache: false }).then(
-      (response) => (this.job_id_csv = response.jobCsv.id)
+      (response) => this.job_id_csv = response.jobCsv.id
     );
 
     this.ecoe_csv = await Job.fetch<Job>(this.job_id_csv, { cache: false });
@@ -156,7 +168,7 @@ export class EcoeResultsComponent implements OnInit {
     }, 500);
   }
   downloadCSV() {
-    this.api.getJobFile(this.job_id_csv, "CSV_" + this.ecoeName);
+    this.api.getJobFile(this.job_id_csv, "csv_" + 'ecoe_' + this.ecoeID);
   }
 
   get show_students(): boolean {
@@ -174,7 +186,7 @@ export class EcoeResultsComponent implements OnInit {
         this.progress_csv = value.progress;
         if (this.progress_csv == 100.0) {
           this.job_csv_file = this.ecoe_csv.uri;
-          this.areGenerated == true;
+          this.areGeneratedCSV = true;
           this.btn_csv = true;
           this.btn_dwl_csv = true;
         }
