@@ -2,15 +2,27 @@
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
+/*
+export const environment = {
+  production: false,
+  TESTING: false,
+  DEBUG: true,
+  API_ROUTE: "http://localhost:5000/backend",
+  CHRONO_ROUTE: "http://localhost:5001",
+  BACK_ROUTE: "http://localhost:5000",
+
+  //   private readonly URI_SOCKETIO = environment.CHRONO_ROUTE;
+};
+*/
+
 export const environment = {
   production: false,
   TESTING: false,
   DEBUG: true,
   API_ROUTE: "http://localhost:8081/backend",
   CHRONO_ROUTE: "http://localhost:8081/chrono",
-  BACK_ROUTE: "http://localhost:8081/",
+  BACK_ROUTE: "http://localhost:8081",
 };
-
 
 /*
  * For easier debugging in development mode, you can import the following file
